requirements/dev.txt
adaptix==3.0.0b4 ; python_version >= "3.11" and python_version < "4.0"
blacksheep==2.0.7 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
guardpost==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
mediatr==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
structlog==24.1.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.27.1 ; python_version >= "3.11" and python_version < "4.0"
nats-py==2.7.2 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.2 ; python_version >= "3.11" and python_version < "4.0"
dishka==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
-r db.txt
